+++ Compilers.slnf
@@ -20,6 +20,8 @@
       "src\\Compilers\\Core\\AnalyzerDriver\\AnalyzerDriver.shproj",
       "src\\Compilers\\Core\\CodeAnalysisTest\\Microsoft.CodeAnalysis.UnitTests.csproj",
       "src\\Compilers\\Core\\MSBuildTaskTests\\Microsoft.Build.Tasks.CodeAnalysis.UnitTests.csproj",
+      "src\\Compilers\\Core\\SdkTaskTests\\Microsoft.Build.Tasks.CodeAnalysis.Sdk.UnitTests.csproj",
+      "src\\Compilers\\Core\\MSBuildTask\\BuildTask.shproj",
       "src\\Compilers\\Core\\MSBuildTask\\MSBuild\\Microsoft.Build.Tasks.CodeAnalysis.csproj",
       "src\\Compilers\\Core\\MSBuildTask\\SDK\\Microsoft.Build.Tasks.CodeAnalysis.Sdk.csproj",
       "src\\Compilers\\Core\\Portable\\Microsoft.CodeAnalysis.csproj",
