+++ src/Compilers/Core/MSBuildTask/MSBuild/Microsoft.Build.Tasks.CodeAnalysis.csproj
@@ -4,31 +4,7 @@
   <PropertyGroup>
     <OutputType>Library</OutputType>
     <RootNamespace>Microsoft.CodeAnalysis.BuildTasks</RootNamespace>
-    <DefaultLanguage>en-US</DefaultLanguage>
     <TargetFrameworks>$(NetRoslynSourceBuild);net472</TargetFrameworks>
-    <AutoGenerateAssemblyVersion>true</AutoGenerateAssemblyVersion>
-
-    <!--
-      Prevent the generation of binding redirects as they are not respected by 
-      MSBuild. At the moment it's on by default due to arcade settings that we
-      need to override here
-
-      https://github.com/dotnet/arcade/issues/9305
-     -->
-    <AutoGenerateBindingRedirects>false</AutoGenerateBindingRedirects>
-    <AssemblyVersion />
-    <!-- CA1819 (Properties should not return arrays) disabled as it is very common across this project. -->
-    <NoWarn>$(NoWarn);CA1819</NoWarn>
-
-    <!-- NuGet -->
-    <IsPackable>true</IsPackable>
-    <PackageId>Microsoft.CodeAnalysis.Build.Tasks</PackageId>
-
-    <PackageDescription>
-      The build task and targets used by MSBuild to run the C# and VB compilers.
-      Supports using VBCSCompiler on Windows.
-    </PackageDescription>
-    <GenerateMicrosoftCodeAnalysisCommitHashAttribute>true</GenerateMicrosoftCodeAnalysisCommitHashAttribute>
   </PropertyGroup>
 
   <ItemGroup>
