+++ src/Compilers/Core/MSBuildTask/Sdk/Microsoft.Build.Tasks.CodeAnalysis.Sdk.csproj
@@ -4,31 +4,8 @@
   <PropertyGroup>
     <OutputType>Library</OutputType>
     <RootNamespace>Microsoft.CodeAnalysis.BuildTasks</RootNamespace>
-    <DefaultLanguage>en-US</DefaultLanguage>
-    <TargetFramework>net472</TargetFramework>
-    <AutoGenerateAssemblyVersion>true</AutoGenerateAssemblyVersion>
-
-    <!--
-      Prevent the generation of binding redirects as they are not respected by 
-      MSBuild. At the moment it's on by default due to arcade settings that we
-      need to override here
-
-      https://github.com/dotnet/arcade/issues/9305
-     -->
-    <AutoGenerateBindingRedirects>false</AutoGenerateBindingRedirects>
-    <AssemblyVersion />
-    <!-- CA1819 (Properties should not return arrays) disabled as it is very common across this project. -->
-    <NoWarn>$(NoWarn);CA1819</NoWarn>
-
-    <!-- NuGet -->
-    <IsPackable>true</IsPackable>
-    <PackageId>Microsoft.CodeAnalysis.Build.Tasks</PackageId>
-
-    <PackageDescription>
-      The build task and targets used by MSBuild to run the C# and VB compilers.
-      Supports using VBCSCompiler on Windows.
-    </PackageDescription>
-    <GenerateMicrosoftCodeAnalysisCommitHashAttribute>true</GenerateMicrosoftCodeAnalysisCommitHashAttribute>
+    <TargetFrameworks>$(NetRoslynSourceBuild);net472</TargetFrameworks>
+    <DefineConstants>SDK_TASK;$(DefineConstants)</DefineConstants>
   </PropertyGroup>
 
   <ItemGroup>
@@ -36,7 +13,7 @@
     <PackageReference Include="Microsoft.Build.Tasks.Core" ExcludeAssets="Runtime" />
     <PackageReference Include="System.Memory" />
     <PackageReference Include="System.Runtime.CompilerServices.Unsafe" />
-    <InternalsVisibleTo Include="Microsoft.Build.Tasks.CodeAnalysis.UnitTests" />
+    <InternalsVisibleTo Include="Microsoft.Build.Tasks.CodeAnalysis.Sdk.UnitTests" />
   </ItemGroup>
 
   <Import Project="..\..\..\..\Dependencies\Contracts\Microsoft.CodeAnalysis.Contracts.projitems" Label="Shared" />
