+++ src/Compilers/Core/MSBuildTaskTests/CscTests.cs
@@ -649,107 +649,5 @@ namespace Microsoft.CodeAnalysis.BuildTasks.UnitTests
                 Assert.Throws<ArgumentException>(() => csc.GenerateResponseFileContents());
             }
         }
-
-        [Fact]
-        public void PathToManagedTool_Normal()
-        {
-            var taskPath = Path.GetDirectoryName(typeof(ManagedCompiler).Assembly.Location)!;
-            var relativePath = RuntimeHostInfo.IsCoreClrRuntime
-                ? Path.Combine("bincore", "csc.dll")
-                : "csc.exe";
-            var task = new Csc();
-            Assert.Equal(Path.Combine(taskPath, relativePath), task.PathToBuiltInTool);
-        }
-
-#if NETFRAMEWORK
-
-        [Fact]
-        public void PathToManagedTool_Bridge()
-        {
-            var taskPath = Path.GetDirectoryName(typeof(ManagedCompiler).Assembly.Location)!;
-            var task = new Csc()
-            {
-                IsSdkFrameworkToCoreBridgeTask = true
-            };
-            Assert.Equal(Path.Combine(taskPath, "..", "bincore", "csc.dll"), task.PathToBuiltInTool);
-        }
-
-#endif
-
-        [Fact]
-        public void IsManagedToolRunningOnCoreClr_Normal()
-        {
-            var task = new Csc();
-            Assert.Equal(RuntimeHostInfo.IsCoreClrRuntime, task.IsBuiltinToolRunningOnCoreClr);
-        }
-
-        [Fact]
-        public void IsManagedToolRunningOnCoreClr_Bridge()
-        {
-#if NET
-            Assert.False(ManagedToolTask.CalculateIsSdkFrameworkToCoreBridgeTask());
-#else
-            var task = new Csc();
-            Assert.False(task.IsBuiltinToolRunningOnCoreClr);
-#endif
-        }
-
-#if NETFRAMEWORK && DEBUG
-
-        [Theory]
-        [InlineData("binfx", true, true)]
-        [InlineData("binfx", false, false)]
-        [InlineData("other", true, false)]
-        [InlineData("other", false, false)]
-        public void CalculateIsSdkFrameworkToCoreBridgeTask_DirectoryName(string dirName, bool makeBincore, bool expected)
-        {
-            LoadInAppDomain(dirName, (appDomain, taskAssemblyName, dirPath) =>
-            {
-                if (makeBincore)
-                {
-                    _ = Directory.CreateDirectory(Path.Combine(Path.GetDirectoryName(dirPath), "bincore"));
-                }
-
-                var testHost = (TaskTestHost)appDomain.CreateInstanceAndUnwrap(taskAssemblyName, typeof(TaskTestHost).FullName);
-                Assert.Equal(expected, testHost.IsSdkFrameworkToCoreBridgeTask);
-            });
-        }
-
-        [Fact]
-        public void CalculateIsSdkFrameworkToCoreBridgeTask_CscPresent()
-        {
-            LoadInAppDomain("binfx", (appDomain, taskAssemblyName, dirPath) =>
-            {
-                File.WriteAllText(Path.Combine(dirPath, "csc.exe"), "real code");
-                var testHost = (TaskTestHost)appDomain.CreateInstanceAndUnwrap(taskAssemblyName, typeof(TaskTestHost).FullName);
-                Assert.False(testHost.IsSdkFrameworkToCoreBridgeTask);
-            });
-        }
-
-        private static void LoadInAppDomain(string dirName, Action<AppDomain, string, string> action)
-        {
-            using var tempRoot = new TempRoot();
-            var dirPath = tempRoot.CreateDirectory().CreateDirectory(dirName).Path;
-            var taskAssembly = typeof(ManagedCompiler).Assembly;
-            var taskFilePath = taskAssembly.Location!;
-            var taskPath = Path.GetDirectoryName(taskFilePath);
-            foreach (var dllPath in Directory.EnumerateFiles(taskPath, "*.dll"))
-            {
-                File.Copy(dllPath, Path.Combine(dirPath, Path.GetFileName(dllPath)));
-            }
-
-            var appDomain = Roslyn.Test.Utilities.Desktop.AppDomainUtils.Create("TestAppDomain", dirPath);
-            try
-            {
-                action(appDomain, taskAssembly.FullName, dirPath);
-            }
-            finally
-            {
-                AppDomain.Unload(appDomain);
-            }
-        }
-
-#endif
-
     }
 }
