+++ src/NuGet/VS.Tools.Roslyn.Package/VS.Tools.Roslyn.Package.csproj
@@ -24,7 +24,7 @@
     <ProjectReference Include="..\..\Compilers\CSharp\csc\AnyCpu\csc.csproj" PrivateAssets="all"/>
     <ProjectReference Include="..\..\Compilers\VisualBasic\vbc\AnyCpu\vbc.csproj" PrivateAssets="all"/>
     <ProjectReference Include="..\..\Interactive\csi\csi.csproj" PrivateAssets="all"/>
-    <ProjectReference Include="..\..\Compilers\Core\MSBuildTask\Microsoft.Build.Tasks.CodeAnalysis.csproj" PrivateAssets="all"/>
+    <ProjectReference Include="..\..\Compilers\Core\MSBuildTask\MSBuild\Microsoft.Build.Tasks.CodeAnalysis.csproj" PrivateAssets="all"/>
     <ProjectReference Include="..\..\Compilers\Server\VBCSCompiler\AnyCpu\VBCSCompiler.csproj" PrivateAssets="all"/>
     <ProjectReference Include="..\..\Tools\ExternalAccess\RazorCompiler\Microsoft.CodeAnalysis.ExternalAccess.RazorCompiler.csproj" PrivateAssets="all" />
   </ItemGroup>
